+++ src/main/java/com/starkbank/utils/Response.java
@@ -129,6 +129,6 @@ public final class Response {
     }
 
     private static String getUserAgent() {
-        return (userAgentOverride == null) ? "Java-" + System.getProperty("java.version") + "-SDK-2.6.0" : userAgentOverride;
+        return (userAgentOverride == null) ? "Java-" + System.getProperty("java.version") + "-SDK-2.7.0" : userAgentOverride;
     }
 }