+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - Workspace.update() to allow parameter updates
 - Transfer.description property to allow control over corresponding Transaction descriptions
 - Base exception class
+- Missing parameters to Boleto, BrcodePayment and Deposit resources 
 
 ### Fixed
 - special characters in brcodePreview query